emotiontrackeragent/src/emotiontrackeragent/requirements.txt
crewai
litellm
python-dotenv
openai-whisper
